Updated contact num to long

+++ CuringBloodCancer/src/Business/People/Donor.java
@@ -26,7 +26,7 @@ public class Donor {
     private String city;
     private String state;
     private int zipCode;
-    private int contact;
+    private long contact;
     private String emailID;
     private String status;
     private Date  lastDonationDate;
@@ -62,8 +62,8 @@ public class Donor {
 // Getter Setter Functions
     public String getName() {return name;}
     public void setName(String name) {this.name = name;}
-    public int getContact() {return contact;}
-    public void setContact(int contact) {this.contact = contact;}
+    public long getContact() {return contact;}
+    public void setContact(long contact) {this.contact = contact;}
     public String getDonorID() {        return donorID;    }
     public void setDonorID(String donorID) {        this.donorID = donorID;    }
     public Date getDob() {        return dob;    }
